requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.1
flask==3.0.0
flask_cors==3.0.10
flask_wtf==1.2.1
fsspec==2023.12.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper==20231117
regex==2023.12.25
requests==2.31.0
sympy==1.12
tiktoken==0.5.2
torch==2.1.2
tqdm==4.66.1
typing-extensions==4.9.0
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0